Refactor component structure: move Header and Footer to commons directory, reintroduce SearchInput component
