+++ src/components/Layout.tsx
@@ -1,11 +1,12 @@
-import Header from './Header';
+import Footer from './commons/Footer';
+import Header from './commons/Header';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
     return (
         <>
             <Header />
             <main>{children}</main>
-            <footer>footer</footer>
+            <Footer />
         </>
     );
 }
